# api/openapi/openapi.json
{
  "components": {
...
    "schemas": {
...
      "CreateScheduleRequest": {
        "properties": {
          "catch_up": {
            "description": "Defaults to `fire-once`.",
            "enum": [
              "fire-once",
              "skip"
            ],
            "type": "string"
          },
          "emitter": {
            "description": "Defaults to `schedule:\u003cname\u003e` when omitted.",
            "type": "string"
          },
          "enabled": {
            "description": "Defaults to true. An operator who declared a cadence and did not say otherwise meant it to run.",
            "type": "boolean"
          },
          "event_name": {
            "minLength": 1,
            "type": "string"
          },
          "first_fire_at": {
            "description": "The first instant this schedule is due, and the phase every later occurrence aligns to. Omitted means one interval from now: creating a schedule is not the same act as starting a run.",
            "format": "date-time",
            "type": "string"
          },
          "interval_seconds": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "payload": {
            "description": "Defaults to `{}`."
          }
        },
        "required": [
          "name",
          "event_name",
          "interval_seconds"
        ],
        "type": "object"
      },
...
      "Error": {
        "description": "The stable error shape, every non-2xx response, every operation. Mirrors internal/clifmt.CliError's JSON encoding exactly: `code` follows the same exit-code-style bucket (1 = domain/user error such as not-found or conflict, 2 = environment error such as a database outage), and the HTTP status line carries the specific condition.",
        "properties": {
          "code": {
            "enum": [
              1,
              2
            ],
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "remediation": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "remediation"
        ],
        "type": "object"
      },
...
      "PatchScheduleRequest": {
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "type": "object"
      },
...
      "Schedule": {
        "properties": {
          "catch_up": {
            "description": "What to do about an occurrence that came due while nothing was running — the question a cron line never answers.\n\n`fire-once` (the default) fires late, exactly once, then realigns to the next declared boundary; missed occurrences are counted, never backfilled, because six queued runs would each carry the same freshly-read findings. `skip` declines a MISSED occurrence and waits for the next boundary, which is right for anything whose value is tied to when it happens — a nightly report has no business running at noon. A skip increments `skip_count` and emits `dev.culture.nodes.schedule.skipped`, so \"chose not to run\" and \"never came due\" stay distinguishable.\n\nThe two differ ONLY for a missed occurrence. One that is merely LATE — overdue by less than one whole interval, which a tick interval and a busy loop produce normally — fires under both.",
            "enum": [
              "fire-once",
              "skip"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "emitter": {
            "description": "Free text naming who emitted the fact. Attribution for operators, never an authority claim (PRD §10.4).",
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "event_name": {
            "description": "The signal event this schedule appends on each occurrence.",
            "type": "string"
          },
          "fire_count": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "interval_seconds": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "last_event_id": {
            "description": "The `signal_events` fact the last fire appended — the audit trail in one field, since the runs it triggered are reachable from it.",
            "type": "string"
          },
          "last_failure_detail": {
            "description": "The repeated reason, verbatim: the last minted run's last attempt's `result.error.detail` (falling back to that attempt's technical status, then the run's, when the attempt reported no error envelope). Absent when the schedule is not currently failing, so its presence is the answer to \"is this schedule holding back, and why\".",
            "type": "string"
          },
          "last_fired_at": {
            "description": "Absent until the first fire.",
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "description": "Unique per namespace. It is the handle an operator disables by, which is why two schedules cannot share one.",
            "type": "string"
          },
          "next_fire_at": {
            "description": "The next instant this schedule is due, and the whole durable state of the loop: it is advanced only inside the transaction that also appends the event, which is what makes a restart unable to double-start or lose an occurrence.",
            "format": "date-time",
            "type": "string"
          },
          "payload": {
            "description": "The event body, verbatim. Plain declared data — a schedule reads no external system and holds no credential."
          },
          "skip_count": {
            "description": "Occurrences declined under a `skip` catch-up policy.",
            "format": "int64",
            "type": "integer"
          },
          "suppressed_count": {
            "description": "Due occurrences this schedule held back because its last two minted runs failed with the SAME reason (task t9, issue #253).\n\nSuppression is a rate, not a stop: a suppressed schedule still mints a probe run at most every `NODES_SCHEDULE_PROBE_INTERVAL` (default 30m), so a repaired environment is discovered without an operator resuming anything, and one completed run puts the schedule straight back on its declared cadence. `enabled` therefore keeps its own meaning — an operator said stop — while this counter records the environment saying stop.\n\nCounted rather than inferred, and separate from `skip_count`, because \"the operator declared a catch-up policy\" and \"the work keeps failing identically\" are different answers to \"why has nothing run\".",
            "format": "int64",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "event_name",
          "emitter",
          "payload",
          "interval_seconds",
          "catch_up",
          "enabled",
          "next_fire_at",
          "fire_count",
          "skip_count",
          "suppressed_count",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "ScheduleList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Schedule"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "actorRegistrationBearerAuth": {
        "description": "The shared secret configured as NODES_ACTOR_REGISTRATION_TOKEN_SECRET. Used only by POST /v1alpha1/actors (task t13) — deliberately its own secret, not the decision secret, so an operator can grant registration standing without also granting the power to decide human tasks.",
        "scheme": "bearer",
        "type": "http"
      },
      "adhocRunBearerAuth": {
        "description": "The shared secret configured as NODES_ADHOC_RUN_TOKEN_SECRET. Used only by POST /v1alpha1/adhoc-runs (task t19, gated by the t15 auth-hardening pass, spec c27) — its own secret, so the power to start ad-hoc (often billable) work is granted independently of decision, registration, and event standing.",
        "scheme": "bearer",
        "type": "http"
      },
      "attemptCallbackBearerAuth": {
        "description": "The short-lived callback token minted for exactly one attempt. The route verifies it against the attempt id in the path before loading durable invocation authority or reading the request body.",
        "scheme": "bearer",
        "type": "http"
      },
      "decisionBearerAuth": {
        "description": "The shared secret configured as NODES_HUMAN_DECISION_TOKEN_SECRET. Used by the three operations that record a human decision — POST /v1alpha1/human-tasks/{id}/decision and the two review operations (POST /v1alpha1/runs/{id}/reviews, POST /v1alpha1/reviews/{id}/commit) — because all three write human-authority records into the ledger on whoever presents the token. Every other operation in this API is authless by phase-1 design (spec decision c45).",
        "scheme": "bearer",
        "type": "http"
      },
      "eventBearerAuth": {
        "description": "The shared secret configured as NODES_EVENT_TOKEN_SECRET. Used only by POST /v1alpha1/events (task t10) — its own secret again, so an external system granted the standing to emit signal events (which can resume any signal-parked run) holds neither decision nor registration power.",
        "scheme": "bearer",
        "type": "http"
      },
      "inboundActorKeyAuth": {
        "description": "Names the party whose credential is being presented — an actor key (`namespace/name`), which is also the mailbox partition the poll claims from and the completion is scoped to. Required together with inboundDialBearerAuth: an absent or empty header is a 401 before the credential is examined at all. The transport routes hardcode the party KIND to `actor`, so a `host`-kind credential cannot dial in even though issuance will mint one.",
        "in": "header",
        "name": "X-Culture-Nodes-Actor-Key",
        "type": "apiKey"
      },
      "inboundDialBearerAuth": {
        "description": "The credential the control plane ISSUED for one dialling party — the `cnd_`-prefixed value returned exactly once by POST /v1alpha1/inbound/credentials. Presented by a bridge on every poll and every completion, alongside inboundActorKeyAuth; neither header is sufficient alone. Verified against a SHA-256 verifier read from PostgreSQL per call, never cached, which is what makes revocation effective on the next dial.",
        "scheme": "bearer",
        "type": "http"
      },
      "inboundIssuanceBearerAuth": {
        "description": "The shared secret configured as NODES_INBOUND_ISSUANCE_TOKEN_SECRET. Used only by POST /v1alpha1/inbound/credentials and POST /v1alpha1/inbound/credentials/revoke (issue #111's dial-in half) — its own secret again, so the standing to mint or end a bridge's IDENTITY is granted independently of decision, registration, event and ad-hoc power. Absent configuration leaves both routes refusing with 401 rather than mounted-but-authless.",
        "scheme": "bearer",
        "type": "http"
...
  },
  "info": {
    "description": "The Culture Nodes control-plane API (docs/initial-design/culture-nodes-prd-spec.md §12.1, §15.1): workflow publication, run orchestration, the append-only work ledger, and review transactions. group: nodes.culture.dev/v1alpha1.\nPhase 1 runs authless by design (spec decision c45): this API carries no authentication or authorization of its own and is meant to run behind a private network. Every operation is reachable by anyone who can reach the listener. Do not expose this API on a public network without a fronting proxy that adds authentication.\nEvery error response uses the Error schema below — the same {code, message, remediation} shape internal/clifmt.CliError renders for the CLI — with the HTTP status line carrying the specific condition. The two reverse-transport routes are the documented exception: POST /v1alpha1/inbound/poll and POST /v1alpha1/inbound/{id}/complete are not mounted through the wrapper that renders Error, so their 400/401 refusals are `text/plain` while their 500s still carry Error. Each of those responses says so where it is declared.",
    "license": {
      "name": "Apache-2.0"
    },
    "summary": "Culture Nodes control-plane API",
    "title": "nodes.culture.dev/v1alpha1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1alpha1/schedules": {
      "get": {
        "description": "Disabled schedules are included deliberately: a schedule that is not running is precisely what an operator asking \"why has nothing happened since Tuesday\" is looking for.",
        "operationId": "listSchedules",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleList"
                }
              }
            },
            "description": "The namespace's schedules."
          }
        },
        "summary": "Every declared schedule, enabled or not.",
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "description": "Task t33 (issue #107): a schedule is an emitter with a clock. On each declared occurrence it appends its declared payload to `signal_events` as a fact, and every published workflow whose `spec.triggers` matches that fact creates a run — the same path a POST /v1alpha1/events delivery takes, so there is exactly one route from \"something happened\" to \"a run exists\".\n\nA schedule is deployment configuration, not part of a workflow: a workflow version is immutable and content-addressed, so a cadence declared inside one could not be disabled without republishing (which would change what the graph IS in order to change how often it runs). It also discovers nothing and holds no credential — what it emits is what an operator wrote down.",
        "operationId": "createSchedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            },
            "description": "The declared schedule."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A declaration this control plane could not honour."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A schedule with that name already exists in this namespace."
          }
        },
        "summary": "Declare a cadence that emits an event with no human action.",
        "tags": [
          "schedules"
        ]
      }
    },
    "/v1alpha1/schedules/{id}": {
      "delete": {
        "description": "The `signal_events` facts it appended and the runs those started are untouched: they are history, and history does not stop being true because the thing that produced it was retired.",
        "operationId": "deleteSchedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such schedule in this namespace."
          }
        },
        "summary": "Remove a schedule.",
        "tags": [
          "schedules"
        ]
      },
      "get": {
        "operationId": "getSchedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            },
            "description": "The schedule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such schedule in this namespace."
          }
        },
        "summary": "One schedule, with its cursor and history.",
        "tags": [
          "schedules"
        ]
      },
      "patch": {
        "description": "`enabled` is the only patchable field. Changing the cadence or the payload of a live schedule would leave already-fired runs pointing at a declaration that no longer exists anywhere, so those are a delete and a create.\n\nDisabling does NOT move `next_fire_at`. A schedule disabled at 09:00 and re-enabled at 17:00 is overdue on re-enable, and its declared `catch_up` policy decides whether that overdue occurrence fires — the same question, with the same answer, as a control plane that was down for those eight hours.",
        "operationId": "patchSchedule",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            },
            "description": "The updated schedule."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No patchable field was supplied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such schedule in this namespace."
          }
        },
        "summary": "Enable or disable a schedule.",
        "tags": [
          "schedules"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "The API is served relative to whatever host and port `nodes serve` binds.",
      "url": "/"
    }
  ],
...
}